Raise error on sqlite3 gem inclusion
When sqlite3 is included in the Gemfile heroku will attempt to install it which fails. This failure is due to a missing sqlite3 library on the dyno which was done intentionally. If we know the build will fail, we should take the opportunity to educate the developer exactly why they should not use sqlite3 on our platform. This is done by directing them to:

https://devcenter.heroku.com/articles/sqlite3

The full output they will see will look like this:

```
Fetching buildpack... done
Detecting buildpack... done, Ruby/Rack
Fetching cache... empty
Compiling app...
  Using Ruby version: ruby-1.9.3
    Cannot install sqlite3 gem on Heroku, remove or replace it with the 'pg' gem
    for more information see: https://devcenter.heroku.com/articles/sqlite3
/tmp/buildpack_Fi0ju/lib/language_pack/ruby.rb:386:in `raise_on_bad_gems!': Cannot install sqlite3 gem (RuntimeError)
	from /tmp/buildpack_Fi0ju/lib/language_pack/ruby.rb:393:in `block in build_bundler'
	from /tmp/buildpack_Fi0ju/lib/language_pack/base.rb:
```

cc @wycats